[MINOR] Fix static code analysis warnings, part II
This patch applies additional changes from an automated static code
analysis. These changes include a number of minor bugs in tests and
system code. The most important bug fix, is the inactive IPA early
abort feature (avoiding unnecessary compilation overhead).

+++ src/main/java/org/apache/sysds/runtime/frame/data/columns/ArrayFactory.java
@@ -91,7 +91,7 @@ public interface ArrayFactory {
 				case STRING:
 					// cannot be known since strings have dynamic length
 					// lets assume something large to make it somewhat safe.
-					return Array.baseMemoryCost() + (long) MemoryEstimates.stringCost(12) * _numRows;
+					return Array.baseMemoryCost() + MemoryEstimates.stringCost(12) * _numRows;
 				default: // not applicable
 					throw new DMLRuntimeException("Invalid type to estimate size of :" + type);
 			}
@@ -116,7 +116,7 @@ public interface ArrayFactory {
 				case STRING:
 					// cannot be known since strings have dynamic length
 					// lets assume something large to make it somewhat safe.
-					return Array.baseMemoryCost() + (long) MemoryEstimates.stringCost(12) * _numRows;
+					return Array.baseMemoryCost() + MemoryEstimates.stringCost(12) * _numRows;
 				case CHARACTER:
 					return Array.baseMemoryCost() + (long) MemoryEstimates.charArrayCost(_numRows);
 				default: // not applicable
